Update xapi-idl to v1.90.1 trunk_ppx version
Also, update the relevant parts of the opam file to match the one in the
source repository at the correct tag, but keep the ppx-specific and more
up-to-date version constraints.

xapi-idl (or xcp-idl) has been rebased on top of master on the ppx_trunk
branch, and a new v1.90.1 version was released on that branch.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>