Remove optcomp
optcomp is a camlp4 syntax extension (but it also works as an
executable). According to opam-repo-dev, only xenopsd used optcomp, but
we updated it to use cppo instead, so we can remove optcomp from this
repository.

It isn't needed by any of the packages here either, because they all use
ppx_optcomp.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>