Simplify and fix Opam setup in Travis
This simplifies the setup of Opam in the Travis build. The Opam
repository only contains libraries from this repo and never uses
the public remote repository.

Previously the OACML_VERSION variable was not effective because it was
not exported and the script was executed in a sub shell. By exporting
it, it is now effective. Likewise, we use OCAML_INIT to avoid having the
Opam repository being initialised outside the travis.sh script.

Signed-off-by: Christian <christian.lindig@citrix.com>