xen-api-client: update to 0.9.16
This fixes a connection leak in the Lwt client that also causes file
descriptors to be leaked in xapi. In the extreme case this bug could
paralyse xapi by causing it to run out of file descriptors in case of a
client that makes many XenAPI calls using the Lwt version of the
xen-api-client library.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>