CA-272542: avoid illegal instructions from nocrypto
Nocrypto detects CPU features at *build time*, instead of runtime.
If the build machine is newer than the machine you install the binary on then it
would crash with an illegal instruction on startup.
See https://github.com/mirleft/ocaml-nocrypto#illegal-instructions and https://github.com/mirleft/ocaml-nocrypto/issues/73

Since there is no runtime detection the only safe thing to do is to disable the
use of these instructions always.
For version 0.5.3 this is done with `--disable-modernity`, for 0.5.4 it'll be
`NOCRYPTO_ACCELERATE=false` environment variable.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>