accept test-only dependencies of xs-extra-dummy as used
Cannot use 'opam install -t' because it tries to install the test
dependencies recursively and fail because we do not define all those
packages.
Opam 2.0 would have a way to avoid that, but with 1.2 we'd have to
resort to parsing a json output to figure out what the first level test
dependencies would be.

Pretend that alcotest is a build dependency instead of a test
dependency. Since we always runs tests during build for xapi-clusterd
this should be equivalent in our use case.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>