add depext to upstream-extra
upstream has switched to OPAM 2 containers by default, but was missing the depext plugin.
This is fine when using the upstream opam repository because it just be
installed, but when we use base remote set to xs-opam this will fail
because it can't find it.
Meantime this got fixed upstream, but nevertheless it would be useful to
have this package because we use it.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>