CA-341597: add conf-libev
Lwt by default chooses a `select` based backend which is limited to 1024
FDs.
There is also a virtual poll implementation that one could hook up to a
`poll()` called via C stubs.

However if `libev` is available then it can use an `epoll` based engine,
which doesn't require any manual C bindings.

Although varstored-guard is the package that needs this, it is
beneficial for other Lwt-based daemons too, so I add the dependency on
xs-toolstack here.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>