Move ocamlformat to upstream/ so we build it by default
It is available in `upstream-extra`, but `opam` doesn't work inside a
pxc/mock chroot anymore because it doesn't have permission to install to
/usr (only an rpm package has).
With planex-buildenv it used to be possible to run things as root and
install additional opam packages.

Easiest is to move this to upstream/ so that xs-opam would build it and
make it available inside build containers.

That way contributors can have matching ocamlformat version to the one
that we use.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>