ci: package and publish tarball to github on tag
Automate tarball and license generation when a tag is pushed

This allows the manual steps of ingesting both of those files to the rpm
build system to be simpler, even automated.

The release is marked as draft, in case the generation goes wrong and we
need to pull it out.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>