-mdx
The current version of 'mdx' is uninstallable with the package set in
xs-opam.
Updating 'mdx' to a newer version would make it installable, but it
pulls in a new 'odoc' which makes 'ocamlformat' uninstallable.

'mdx' is unused, so drop it for now. We can revisit this again during
the 4.14 update.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>