packages: work around opam bugs
When producing the cache, opam for some packages does not create files
for all the hashes.

When installing the packages, opam only tries the first hash before
trying to download a package, not all of the hashes.

The combination of both makes makes the rpm builds to fail when these
are properly isolated.

I've seen this happen only for these two packages.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>