upstream-extra: add ocaml-print-intf.1.2.0
This can be used to create a `.mli` file (or at least a starting point for one)
when `dune` is used and you have a lot of dependencies (so that invoking `ocamlc` by hand would be too tedious).

It is recommended by the dune docs:
https://dune.readthedocs.io/en/latest/faq.html#how-can-i-generate-an-mli-file-from-an-ml-file

I had this pinned in my local opam switch for a long time.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>