upstream-extra: add opam-dune-lint.0.6 and dependencies
This can be used to keep `dune-project` and `*.opam.template` files up-to-date
by finding dependencies that are declared in `dune` files, but not in the `opam` package.
It doesn't yet find all such bugs (https://github.com/ocurrent/opam-dune-lint/issues/71),
but it is a starting point.

It also adds its dependencies for parsing opam files:
* ocamlgraph.2.1.0 (I'm surprised we didn't already have this)
* opam-core, opam-file-format, opam-format, opam-repository-opam-state 2.2.1
* spdx_licenses.1.2.0, swhid_core.0.1

I had this pinned in my local opam repo for a while too, this is the latest version.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>