tools: show more comprehensible output for sync-xapi-metadata
Because opam metadata can be either on the root directory, or on opam/,
there may be up to 2 requests per package. The first one may fail, but
the second one can't.

Supress all stderr output for the first command, and supress progress
output for the second one, but not actual errors.

Also print the package name before trying to download the metadata to
know which package failed to be downloaded.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>