packages: replace ocaml 5.3 with ocaml 5.4
Updates ppxlib so it supports the new compiler AST. Unfortunately JST
does not support old compilers and new versions of their libraries need
to be brought in while keeping the existing ones. The same thing happens
with rpclibs.

Similarly with the new lsp server: the new merlin and lsp libraries
require the new ocaml compiler.

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>